// src/ui-kit/components/modal-old/modal-layout/types.ts
export declare namespace ModalDismissSource {
  /**
   * Description of dismiss source. For now this value only purpose is for analytics.
   *  **So avoid using this field anywhere else!**
   *
   * @example "my fancy button was clicked:3"
   */
  // eslint-disable-next-line sonar/redundant-type-aliases
  export type Description = string;

  export type Input = 'unknown' | 'ui' | 'hardware';
  export type Origin = 'unknown' | 'user' | 'external';
}

export type ModalDismissSource =
  | {
      origin: ExtractStrict<ModalDismissSource.Origin, 'unknown'>;
      input: ExtractStrict<ModalDismissSource.Input, 'unknown'>;
      description: ModalDismissSource.Description | undefined;
    }
  | {
      origin: ExtractStrict<ModalDismissSource.Origin, 'external'>;
      input: ExtractStrict<ModalDismissSource.Input, 'unknown'>;
      description: ModalDismissSource.Description | undefined;
    }
  | {
      origin: ExtractStrict<ModalDismissSource.Origin, 'user'>;
      input: ModalDismissSource.Input;
      description: ModalDismissSource.Description | undefined;
    };

export namespace ModalDismissAction {
  /**
   * - `cancel` — modal was closed without accessing intended user-flow.
   *   Example: click on backdrop, close request from outside, click on header
   *   close button, click on ESC button and etc.
   * - `resolve` — modal was closed via intended user-flow.
   */
  export type Intent = 'cancel' | 'resolve';
}

export interface ModalDismissAction {
  intent: ModalDismissAction.Intent;
  source: ModalDismissSource;
}
// ...
export type InferDataTypeFromModalComponent<TComponent extends Component> =
  CustomComponentProps<TComponent> extends { data: infer DataType } ? DataType
  : never;

export type InferValueTypeFromModalComponent<TComponent extends Component> =
  CustomComponentProps<TComponent> extends { modelValue: infer ValueType } ?
    ValueType
  : never;